requirements.txt

streamlit
python-docx
PyMuPDF
supabase
